credits update again
so; plugged dreams came out and i was very, very intrigued by how they got monsters running around on the arena and everything; and in the process implemented it and the rest of the arena effects suite on top of the existing swordsoul gimmick secretboss i was already working on.
but commit history is only half a credit; folks, you gotta have it be public as well... hence adding it like this.

I hope not to make a habit of just GRABBING parts of other projects like this; it's generally lame + leads to laziness verse a ballooned scope - I wouldn't have done it if i didn't consider it worth it in this instance. Funkin' did an absolutely phenominal job with the bullets and effectswork there, tbh.

moreover i am of the opinion that carthage should be destroyed